routes for leader ready

+++ app/src/main/java/com/example/poapp/model/dao/RouteSectionDAO.kt
@@ -16,7 +16,7 @@ interface RouteSectionDAO {
     fun deleteAll()
 
     @Query("select * from OdcinkiTras where FKtrasa = :routeID order by id asc")
-    fun getRouteSectionForRoute(routeID: Long): List<RouteSection>
+    fun getRouteSectionForRoute(routeID: Int): List<RouteSection>
 
     @Query("delete from OdcinkiTras where FKtrasa=:routeID")
     fun deleteAllFor(routeID: Long)
